[get_d64.c,set_d64.c] small fixes (did not work any more with --with-gmp-build)
[From SVN r8405 (trunk)]

+++ src/set_d64.c
@@ -223,7 +223,7 @@ static void
 decimal64_to_string (char *s, _Decimal64 d)
 {
   int sign = 0, n;
-  long exp = 0;
+  int exp = 0;
   char *s0 = s;
 
   if (d != d) /* NaN */
